requirements.txt
openai==0.27.8
Pillow==10.0.0
Requests==2.31.0
selenium==4.10.0
